Add genetic algorithm template

+++ utils/destination.jl
@@ -1,3 +1,3 @@
-function destination(graph::Matrix{Float64}, road::Array{Int})
+function destination(graph::Matrix{Float64}, road::Array{Int})::Float64
     return sum(graph[road[i], road[i+1]] for i in 1:length(road)-1) + graph[road[length(road)], road[1]]
 end
